% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odum_symbols.R
\name{odum_heat_sink}
\alias{odum_heat_sink}
\title{Odum ESL HEAT SINK — small triangle pointing down.
Represents dispersed (unusable) heat at the end of a transformation.}
\usage{
odum_heat_sink(cx, cy, s, id)
}
\arguments{
\item{cx, cy}{Numeric centre (top vertex of triangle).}

\item{s}{Triangle side.}

\item{id}{Group id.}
}
\description{
Odum ESL HEAT SINK — small triangle pointing down.
Represents dispersed (unusable) heat at the end of a transformation.
}
