% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odum_symbols.R
\name{odum_interaction}
\alias{odum_interaction}
\title{Odum ESL INTERACTION — diamond (multiplicative junction).
Represents where two or more flows combine multiplicatively.}
\usage{
odum_interaction(cx, cy, s, id)
}
\arguments{
\item{cx, cy}{Numeric centre.}

\item{s}{Diamond size (side).}

\item{id}{Group id.}
}
\description{
Odum ESL INTERACTION — diamond (multiplicative junction).
Represents where two or more flows combine multiplicatively.
}
